requirements.txt
mail-parser>=4.1.2
imapclient>=3.1.0
html2text>=2020.1.16
dnspython>=2.0.0
publicsuffix2>=2.20190812
expiringdict==1.2.2
dkimpy>=1.1.0
cryptography>=41.0.0
msgraph-sdk>=1.0.0
azure-identity>=1.15.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
nose
pygments
ruff
doc8
collective.checkdocs
jinja2
packaging
imagesize
alabaster
babel
rstcheck>=3.3.1
docutils<0.19,>=0.7
sphinx
sphinx_rtd_theme
myst-parser
myst-parser[linkify]
wheel
codecov
hatch